Added the language list used
This file contains the order of languages used in Google Translate. It will most likely be used in other Google Translated mods, and you can use it yourself if you want. The file isn't part of the mod itself, and as such, is not necessary to be included (hence why it isn't in a folder)

I can't spell